Move exception to new billing directory
