+++ src/main/java/bio/terra/service/billing/flight/AuthorizeRawlsBillingProjectsUseStep.java
@@ -2,8 +2,8 @@ package bio.terra.service.billing.flight;
 
 import bio.terra.common.iam.AuthenticatedUserRequest;
 import bio.terra.service.job.DefaultUndoStep;
-import bio.terra.service.profile.exception.BillingProjectNotAccessibleException;
 import bio.terra.service.rawls.RawlsService;
+import bio.terra.service.billing.exception.BillingProjectNotAccessibleException;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.StepResult;
 import bio.terra.stairway.StepStatus;
